support updating primary/unique key columns in new DML (#20842)
let DML queries to run in single-step plan

Approved by: @daviszhen, @aressu1985, @zhangxu19830126, @badboynt1, @qingxinhome, @ouyuanning, @XuPeng-SH

+++ pkg/sql/plan/bind_delete.go
@@ -230,7 +230,7 @@ func (builder *QueryBuilder) bindDelete(stmt *tree.Delete, bindCtx *BindContext)
 				lockTarget := &plan.LockTarget{
 					TableId:            tableDef.TblId,
 					ObjRef:             DeepCopyObjectRef(dmlCtx.objRefs[i]),
-					PrimaryColIdxInBat: int32(pkPos),
+					PrimaryColIdxInBat: pkPos,
 					PrimaryColRelPos:   selectNodeTag,
 					PrimaryColTyp:      col.Typ,
 				}
@@ -266,7 +266,7 @@ func (builder *QueryBuilder) bindDelete(stmt *tree.Delete, bindCtx *BindContext)
 						lockTargets = append(lockTargets, &plan.LockTarget{
 							TableId:            idxNode.TableDef.TblId,
 							ObjRef:             DeepCopyObjectRef(idxNode.ObjRef),
-							PrimaryColIdxInBat: int32(pkPos),
+							PrimaryColIdxInBat: pkPos,
 							PrimaryColRelPos:   idxNode.BindingTags[0],
 							PrimaryColTyp:      col.Typ,
 						})
